#36 refactor: 명범이 피드백 반영 (UIEvent -> BillingEvent)
