+++ presentation/src/main/java/com/nextroom/nextroom/presentation/ui/billing/BillingEvent.kt
@@ -1,0 +1,6 @@
+package com.nextroom.nextroom.presentation.ui.billing
+
+sealed interface BillingEvent {
+    data object PurchaseAcknowledged : BillingEvent
+    data class PurchaseFailed(val purchaseState: Int) : BillingEvent
+}
